+++ packages/cross-toolchain/mpfr.sh
@@ -14,4 +14,6 @@ PKG_FILE="$(
 )"
 
 python "$SCRIPTS/download.py" -f "$YAML_FILE"
-mv -v "$LFS_SOURCES/$PKG_FILE" mpfr
+if [[ -f $LFS_SOURCES/$PKG_FILE ]]; then
+  mv -v "$LFS_SOURCES/$PKG_FILE" mpfr
+fi
